+++ epoch1d/src/deck/deck_control_block.F90
@@ -30,7 +30,7 @@ MODULE deck_control_block
   PUBLIC :: control_block_start, control_block_end
   PUBLIC :: control_block_handle_element, control_block_check
 
-  INTEGER, PARAMETER :: control_block_elements = 36 + 4 * c_ndims
+  INTEGER, PARAMETER :: control_block_elements = 37 + 4 * c_ndims
   LOGICAL, DIMENSION(control_block_elements) :: control_block_done
   ! 3rd alias for ionisation
   CHARACTER(LEN=string_length) :: ionization_alias = 'field_ionization'
@@ -77,7 +77,8 @@ MODULE deck_control_block
           'reset_walltime           ', &
           'dlb_maximum_interval     ', &
           'dlb_force_interval       ', &
-          'balance_first            ' /)
+          'balance_first            ', &
+          'use_pre_balance          ' /)
   CHARACTER(LEN=string_length), DIMENSION(control_block_elements) :: &
       alternate_name = (/ &
           'nx                       ', &
@@ -119,7 +120,8 @@ MODULE deck_control_block
           'reset_walltime           ', &
           'balance_maximum_interval ', &
           'balance_force_interval   ', &
-          'balance_first            ' /)
+          'balance_first            ', &
+          'use_pre_balance          ' /)
 
 CONTAINS
 
@@ -142,6 +144,7 @@ CONTAINS
       ic_from_restart = .FALSE.
       neutral_background = .TRUE.
       use_particle_migration = .FALSE.
+      use_pre_balance = .TRUE.
       restart_number = 0
       check_stop_frequency = 10
       stop_at_walltime = -1.0_num
@@ -405,6 +408,8 @@ CONTAINS
       dlb_force_interval = as_integer_print(value, element, errcode)
     CASE(4*c_ndims+36)
       balance_first = as_logical_print(value, element, errcode)
+    CASE(4*c_ndims+37)
+      use_pre_balance = as_logical_print(value, element, errcode)
     END SELECT
 
   END FUNCTION control_block_handle_element