+++ epoch3d/src/housekeeping/setup.F90
@@ -27,6 +27,7 @@ MODULE setup
   USE window
   USE timer
   USE helper
+  USE balance
   USE sdf
 
   IMPLICIT NONE
@@ -36,7 +37,7 @@ MODULE setup
   PUBLIC :: after_control, minimal_init, restart_data
   PUBLIC :: open_files, close_files, flush_stat_file
   PUBLIC :: setup_species, after_deck_last, set_dt
-  PUBLIC :: read_cpu_split
+  PUBLIC :: read_cpu_split, pre_load_balance
 
   TYPE(particle), POINTER, SAVE :: iterator_list
 #ifndef NO_IO
@@ -1866,5 +1867,18 @@ CONTAINS
     window_shift(1) = window_offset
 
   END SUBROUTINE
+
+
+
+  SUBROUTINE pre_load_balance
+
+    IF (.NOT.use_pre_balance .OR. nproc == 1) RETURN
+
+    pre_loading = .TRUE.
+    CALL auto_load
+    CALL pre_balance_workload
+    pre_loading = .FALSE.
+
+  END SUBROUTINE pre_load_balance
 
 END MODULE setup