+++ epoch3d/src/user_interaction/helper.F90
@@ -68,10 +68,22 @@ CONTAINS
 
   SUBROUTINE auto_load
 
-    INTEGER :: ispecies
+    INTEGER :: ispecies, io, iu
     TYPE(particle_species), POINTER :: species
 
     CALL set_thermal_bcs
+
+    IF (pre_loading .AND. n_species > 0) THEN
+      ALLOCATE(npart_per_cell_array(nx,ny,nz))
+      npart_per_cell_array = 0
+    ELSE IF (n_species > 0) THEN
+      IF (rank == 0) THEN
+        DO iu = 1, nio_units ! Print to stdout and to file
+          io = io_units(iu)
+          WRITE(io,*) 'Attempting to load particles'
+        END DO
+      END IF
+    END IF
 
     DO ispecies = 1, n_species
       species => species_list(ispecies)
@@ -87,6 +99,8 @@ CONTAINS
           species_list(ispecies)%initial_conditions%density_min, &
           species_list(ispecies)%initial_conditions%density_max)
 #endif
+      IF (pre_loading) CYCLE
+
       CALL setup_particle_temperature(&
           species_list(ispecies)%initial_conditions%temp(:,:,:,1), c_dir_x, &
           species, species_list(ispecies)%initial_conditions%drift(:,:,:,1))
@@ -97,6 +111,8 @@ CONTAINS
           species_list(ispecies)%initial_conditions%temp(:,:,:,3), c_dir_z, &
           species, species_list(ispecies)%initial_conditions%drift(:,:,:,3))
     END DO
+
+    IF (pre_loading) RETURN
 
     IF (rank == 0) THEN
       DO ispecies = 1, n_species
@@ -215,6 +231,21 @@ CONTAINS
     CALL MPI_ALLREDUCE(density_total, density_total_global, 1, mpireal, &
         MPI_SUM, comm, errcode)
     density_average = density_total_global / REAL(num_valid_cells_global, num)
+
+    IF (pre_loading) THEN
+      DO iz = 1, nz
+      DO iy = 1, ny
+      DO ix = 1, nx
+        npart_per_cell = NINT(density(ix,iy,iz) / density_average &
+            * npart_per_cell_average)
+        npart_per_cell_array(ix,iy,iz) = &
+            npart_per_cell_array(ix,iy,iz) + INT(npart_per_cell)
+      END DO ! ix
+      END DO ! iy
+      END DO ! iz
+
+      RETURN
+    END IF
 
     npart_this_proc_new = 0
     DO iz = 1, nz
@@ -443,6 +474,23 @@ CONTAINS
       species%npart_per_cell = &
           REAL(npart_this_species,num) / num_valid_cells_global
       npart_per_cell = FLOOR(species%npart_per_cell, KIND=i8)
+    END IF
+
+    IF (pre_loading) THEN
+      IF (npart_per_cell <= 0) RETURN
+
+      DO iz = iz_min, iz_max
+      DO iy = iy_min, iy_max
+      DO ix = ix_min, ix_max
+        IF (.NOT. load_list(ix,iy,iz)) CYCLE
+
+        npart_per_cell_array(ix,iy,iz) = &
+            npart_per_cell_array(ix,iy,iz) + INT(npart_per_cell)
+      END DO ! ix
+      END DO ! iy
+      END DO ! iz
+
+      RETURN
     END IF
 
     partlist => species%attached_list